requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
qemu.qmp==0.0.3
requests==2.32.3
ruff==0.4.8
toml==0.10.2
tqdm==4.66.4
urllib3==2.2.1
